+++ BinaryTree/BinaryTree.Lib/BinaryTree.cs
@@ -28,110 +28,18 @@ namespace BinaryTreeEx.Lib
             }
         }
 
-        public bool IsPresent(int v)
-        {
-            return _root.IsPreSent(v);
-        }
+        public bool IsPresent(int v) { return _root.IsPreSent(v); }
         public int Depth { get { return _root.Depth; } }
         public int Count { get { return _root.Count; } }
         public int Sum { get { return _root.Value + _root.Sum; } }
-
-        public void InOrderOutput()
+        public int LCommonAncestor(int v1, int v2)
         {
-            _root.InOrderTraversal();
+            return _root.LCommonAncestor(Math.Min(v1, v2), Math.Max(v1, v2)).Value;
         }
-
+        public void InOrderOutput() { _root.InOrderTraversal(); }
         public override string ToString()
         {
             return "Count = " + Convert.ToString(Count);
         }
     }
-
-
-    internal class TreeNode
-    { 
-        private int _value;
-        private TreeNode _leftChild, _rightChild;
-        public TreeNode(int v)
-        {
-            _value = v;
-            _leftChild = null;
-            _rightChild = null;
-        }
-
-        public int Value { get { return _value; } }
-
-        internal void Insert(int v)
-        {
-            if(v < _value)
-            {
-                if(_leftChild != null)
-                    _leftChild.Insert(v);
-                else
-                    _leftChild = new TreeNode(v);
-            }   
-            else
-            {
-                if(_rightChild != null)
-                    _rightChild.Insert(v);
-                else
-                    _rightChild = new TreeNode(v);
-            }
-        }
-
-        public override string ToString()
-        {
-            return Convert.ToString(_value);
-        }
-
-        internal bool IsPreSent(int v)
-        {
-            if (v == _value)
-                return true;
-            if (v >= _value)
-                return (_rightChild == null) ? false : _rightChild.IsPreSent(v);
-            else
-                return (_leftChild == null) ? false : _leftChild.IsPreSent(v);
-        }
-
-        internal void InOrderTraversal()
-        {
-            if (_leftChild != null)
-                _leftChild.InOrderTraversal();
-            Console.WriteLine(this);
-            if (_rightChild != null)
-                _rightChild.InOrderTraversal();       
-        }
-
-        internal int Sum
-        {
-            get
-            {
-                return ((_leftChild == null) ? 0 : _leftChild._value + _leftChild.Sum)
-                    + ((_rightChild == null) ? 0 : _rightChild._value + _rightChild.Sum);
-            }
-        }
-
-        internal int Count
-        {
-            get
-            {
-                return 1 + ((_leftChild == null) ? 0 : _leftChild.Count)
-                    + ((_rightChild == null) ? 0 : _rightChild.Count);        
-            }
-        }
-
-        internal int Depth
-        {
-            get
-            {
-                if(_leftChild == null & _rightChild == null)
-                    return 1;
-                int leftDepth = (_leftChild == null) ? 0 : _leftChild.Depth;
-                int rightDepth = (_rightChild == null) ? 0 : _rightChild.Depth;
-                return 1 + Math.Max(leftDepth, rightDepth);
-            }
-        }
-    }
-
 }
